updated the comments dropped
